feat: 계좌이체 기능보완
- 거래이력 생성

+++ BackEnd/src/main/java/com/dokb/DoKB/history/domain/History.java
@@ -6,7 +6,6 @@ import lombok.*;
 import lombok.experimental.Accessors;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -18,36 +17,36 @@ import java.time.LocalDateTime;
 @ToString(exclude = {"account"})
 public class History {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    private String opponentAccount;
+	private String opponentAccount;
 
-    private LocalDateTime dealDate;
+	private LocalDateTime dealDate;
 
-    private String inOut;
+	private String inOut;
 
-    private BigDecimal amount;
+	private Long amount;
 
-    private BigDecimal balance;
+	private Long balance;
 
 
-    //History n:1 account
-    @JsonBackReference
-    @ManyToOne
-    @JoinColumn(name = "accountNumber")
-    private Account account;
+	//History n:1 account
+	@JsonBackReference
+	@ManyToOne
+	@JoinColumn(name = "accountNumber")
+	private Account account;
 
-    public HistoryApi parseHistoryApi(){
-        return HistoryApi.builder()
-                .id(this.getId())
-                .opponentAccount(this.getOpponentAccount())
-                .dealDate(this.getDealDate())
-                .inOut(this.getInOut())
-                .amount(this.getAmount())
-                .balance(this.getBalance())
-                .accountNumber(this.getOpponentAccount())
-                .build();
-    }
+	public HistoryApi parseHistoryApi() {
+		return HistoryApi.builder()
+				.id(this.getId())
+				.opponentAccount(this.getOpponentAccount())
+				.dealDate(this.getDealDate())
+				.inOut(this.getInOut())
+				.amount(this.getAmount())
+				.balance(this.getBalance())
+				.accountNumber(this.getOpponentAccount())
+				.build();
+	}
 }
